move font definitions to their own file
